fix: cast object_store file size to usize
- Update AWS and GCP downloaders to cast `meta.size` to `usize`
- Ensure compatibility with expected type in `FileMetadata`

+++ src/cache/file_downloader/impl_object_store/aws.rs
@@ -137,7 +137,7 @@ impl FileDownloader for AwsS3Downloader {
         let path = object_store::path::Path::parse(path)?;
         match self.object_store.head(&path).await {
             Ok(meta) => Ok(FileMetadata {
-                size: meta.size,
+                size: meta.size as usize,
                 location: meta.location.to_string(),
                 last_modified: meta.last_modified,
                 e_tag: meta.e_tag,
